31.homework/requirements.txt
pytest==7.4.0
selenium==4.11.2
webdriver-manager==4.0.0
allure-pytest==2.13.2
pytest-xdist==3.3.1
faker==19.3.0